Refactor Footer into modular subcomponents
Replace the monolithic src/components/Footer.tsx with a new modular implementation at src/components/Footer/Footer.tsx and split UI/logic into four subcomponents: ActiveItemButton, DecibelMeterVisualization, SOSTrigger, and SolarCycleNotification. Update AppShell import to use the new Footer/Footer path and remove the old Footer.tsx file. SOSTrigger now exposes the press state and progress animation to the parent so the footer can render a fuse timer overlay; other behavior (decibel meter, solar notifications, active-item handling) is preserved but organized for improved maintainability.
